+++ simtong-application/src/test/kotlin/team/comit/simtong/user/usecase/SignUpUseCaseTests.kt
@@ -16,8 +16,8 @@ import team.comit.simtong.domain.user.dto.DomainSignUpRequest
 import team.comit.simtong.domain.user.model.Authority
 import team.comit.simtong.domain.user.model.User
 import team.comit.simtong.domain.user.policy.SignUpPolicy
-import team.comit.simtong.domain.user.spi.CheckAuthCodePolicyPort
-import team.comit.simtong.domain.user.spi.CheckUserPort
+import team.comit.simtong.domain.auth.spi.CheckAuthCodePolicyPort
+import team.comit.simtong.domain.auth.spi.CheckUserPort
 import team.comit.simtong.domain.user.spi.SaveUserPort
 import team.comit.simtong.domain.user.spi.SecurityPort
 import team.comit.simtong.domain.user.usecase.SignUpUseCase
